app/requirements.txt
tensorflow==2.19.0
scikit-learn==1.6.1
numpy==2.1.0
matplotlib==3.10.3
seaborn==0.13.2
pandas==2.2.3
streamlit==1.45.0
librosa==0.11.0
Pillow
